Update imports by moving dynamic url into degree directory
